Some cleanup. Moved decorators to separate file.
